Remove the assumption about image resolution
When looking through the image files in the source directory, images that are meant for mobile devices needed to be discarded.
Previously this was done by only taking the images with 1920x1080 resolution, which worked but is potentially an incorrect decision.
With this fix, we determine if an image is for desktop and not for mobile by simply checking if its width > its height, removing the assumption about the screen resolution.